Add known_hosts handling for libssh2 implementation
Add known_hosts handling for libssh2 implementation. known_hosts file is stored in
same place as before, but is now written with the app's creator code ('Ssh2').
If a known_hosts file with 'CWIE' is found, it will be renamed and a new file written
in its place (the old lsh format is not compatible).
SHA256 hostkey hashes are used, and stored in the file by the hostname as entered by the user
(I think this matches the OpenSSH client behavior). Port numbers are not added though.